reddit_video_agent/requirements.txt
google-generativeai
playwright
openai-whisper
moviepy
rembg
srt
pydub
pillow
numpy
requests
python-dotenv
beautifulsoup4
yt-dlp
srt
